+++ src/idunno.Bluesky/Labeler/Defs/LabelerViewerState.cs
@@ -3,7 +3,9 @@
 
 using idunno.AtProto;
 
+#pragma warning disable IDE0130 // Namespace does not match folder structure
 namespace idunno.Bluesky.Labeler;
+#pragma warning restore IDE0130 // Namespace does not match folder structure
 
 /// <summary>
 /// Gets a view of the relationship between the requesting actor of a labeler view, and the labeler.